requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=2.2
whitenoise>=6.7
gunicorn>=21.2

ruff>=0.5
pytest>=8.0
pytest-django>=4.8
